Code 5
Aggregate Functions in SQL combined with Group By Order By and Limit Statements. 
COUNT counts how many rows are in a particular column.
SUM adds together all the values in a particular column.
MIN and MAX return the lowest and highest values in a particular column, respectively.
AVG calculates the average of a group of selected values.